requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.1
django-crispy-forms==1.13.0
django-embed-video==1.4.1
idna==3.3
jsmin==3.0.1
Pillow==9.0.0
pytz==2021.1
requests==2.27.1
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.8